# contracts/v1/openapi.json
{"openapi":"3.1.0","info":{"title":"Staff Server V1 Contract","version":"v1"},"paths":{"/api/client/applications":{"post":{"parameters":[{"$ref":"#/components/parameters/ContentTypeHeader"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["business_name","requested_amount","lender_id","product_id"],"properties":{"business_name":{"type":"string"},"requested_amount":{"type":"number"},"lender_id":{"type":"string"},"product_id":{"type":"string"}}}}}},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"type":"object","required":["application"],"properties":{"application":{"$ref":"#/components/schemas/Application"}}}}}},"400":{"$ref":"#/components/responses/Error"}}}},"/api/client/applications/{id}":{"patch":{"parameters":[{"$ref":"#/components/parameters/ContentTypeHeader"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":false,"properties":{"business_name":{"type":"string"},"requested_amount":{"type":"number"},"metadata":{"type":"object","additionalProperties":true}}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["application"],"properties":{"application":{"$ref":"#/components/schemas/Application"}}}}}},"404":{"$ref":"#/components/responses/Error"}}}},"/api/client/application/{id}/status":{"get":{"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["status"],"properties":{"status":{"type":"object","required":["applicationId","pipelineState","processingStage","updatedAt"],"properties":{"applicationId":{"type":"string"},"pipelineState":{"type":["string","null"]},"processingStage":{"type":["string","null"]},"updatedAt":{"type":"string","format":"date-time"}}}}}}}},"404":{"$ref":"#/components/responses/Error"}}}},"/api/client/submissions":{"post":{"parameters":[{"$ref":"#/components/parameters/ContentTypeHeader"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"201":{"description":"Created","content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"400":{"$ref":"#/components/responses/Error"}}}},"/api/client/documents":{"post":{"parameters":[{"$ref":"#/components/parameters/MultipartHeader"}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","required":["applicationId","category","file"],"properties":{"applicationId":{"type":"string"},"category":{"type":"string"},"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Document"}}}},"400":{"$ref":"#/components/responses/Error"}}}},"/api/client/documents/upload":{"post":{"parameters":[{"$ref":"#/components/parameters/MultipartHeader"}],"requestBody":{"required":true,"content":{"multipart/form-data":{"schema":{"type":"object","required":["applicationId","category","file"],"properties":{"applicationId":{"type":"string"},"category":{"type":"string"},"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Document"}}}},"400":{"$ref":"#/components/responses/Error"}}}},"/api/client/documents/{id}/presign":{"get":{"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"404":{"$ref":"#/components/responses/Error"}}}},"/api/client/documents/{id}/accept":{"post":{"parameters":[{"$ref":"#/components/parameters/AuthHeader"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["ok"],"properties":{"ok":{"type":"boolean"}}}}}},"401":{"$ref":"#/components/responses/Error"}}}},"/api/client/documents/{id}/reject":{"post":{"parameters":[{"$ref":"#/components/parameters/AuthHeader"},{"$ref":"#/components/parameters/ContentTypeHeader"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["reason"],"properties":{"reason":{"type":"string"}}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["ok"],"properties":{"ok":{"type":"boolean"}}}}}},"401":{"$ref":"#/components/responses/Error"}}}},"/api/client/lender-products":{"get":{"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","additionalProperties":true}}}}}}}},"/api/client/lender-products/{id}/requirements":{"get":{"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}}}}},"/api/client/lender-products/requirements":{"get":{"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}}}}},"/api/client/lenders":{"get":{"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Lender"}}}}}}}},"/api/portal/applications":{"get":{"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["items"],"properties":{"items":{"type":"array","items":{"type":"object","additionalProperties":true}}}}}}}}}},"/api/portal/applications/stages":{"get":{"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}}}}},"/api/portal/applications/{id}":{"get":{"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["application","pipeline","documents"],"properties":{"application":{"$ref":"#/components/schemas/Application"},"pipeline":{"type":"object","additionalProperties":true},"documents":{"type":"array","items":{"type":"object","additionalProperties":true}}}}}}},"404":{"$ref":"#/components/responses/Error"}}}},"/api/portal/applications/{id}/history":{"get":{"parameters":[{"$ref":"#/components/parameters/AuthHeader"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["items"],"properties":{"items":{"type":"array","items":{"type":"object","additionalProperties":true}}}}}}},"401":{"$ref":"#/components/responses/Error"}}}},"/api/portal/applications/{id}/status":{"patch":{"parameters":[{"$ref":"#/components/parameters/AuthHeader"},{"$ref":"#/components/parameters/ContentTypeHeader"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["status"],"properties":{"status":{"type":"string"},"reason":{"type":["string","null"]}}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["ok","applicationId","status"],"properties":{"ok":{"type":"boolean"},"applicationId":{"type":"string"},"status":{"type":"string"}}}}}},"400":{"$ref":"#/components/responses/Error"}}}},"/api/portal/applications/{id}/promote":{"post":{"parameters":[{"$ref":"#/components/parameters/AuthHeader"},{"$ref":"#/components/parameters/ContentTypeHeader"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["reason"],"properties":{"reason":{"type":"string"},"nextStage":{"type":"string"}}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["ok","applicationId","nextStage"],"properties":{"ok":{"type":"boolean"},"applicationId":{"type":"string"},"nextStage":{"type":"string"}}}}}},"400":{"$ref":"#/components/responses/Error"}}}},"/api/portal/applications/{id}/retry":{"post":{"parameters":[{"$ref":"#/components/parameters/AuthHeader"},{"$ref":"#/components/parameters/ContentTypeHeader"}],"requestBody":{"required":false,"content":{"application/json":{"schema":{"type":"object","properties":{"reason":{"type":"string"}}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["job"],"properties":{"job":{"type":"object","additionalProperties":true}}}}}}}}},"/api/portal/applications/{id}/retry-job":{"post":{"parameters":[{"$ref":"#/components/parameters/AuthHeader"},{"$ref":"#/components/parameters/ContentTypeHeader"}],"requestBody":{"required":false,"content":{"application/json":{"schema":{"type":"object","properties":{"reason":{"type":"string"}}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["job"],"properties":{"job":{"type":"object","additionalProperties":true}}}}}}}}},"/api/portal/jobs/{id}/history":{"get":{"parameters":[{"$ref":"#/components/parameters/AuthHeader"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["items"],"properties":{"items":{"type":"array","items":{"type":"object","additionalProperties":true}}}}}}}}}},"/api/portal/jobs/{id}/retry":{"post":{"parameters":[{"$ref":"#/components/parameters/AuthHeader"},{"$ref":"#/components/parameters/ContentTypeHeader"}],"requestBody":{"required":false,"content":{"application/json":{"schema":{"type":"object","properties":{"reason":{"type":"string"}}}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["job"],"properties":{"job":{"type":"object","additionalProperties":true}}}}}}}}},"/api/portal/documents/{id}/history":{"get":{"parameters":[{"$ref":"#/components/parameters/AuthHeader"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["items"],"properties":{"items":{"type":"array","items":{"type":"object","additionalProperties":true}}}}}}}}}},"/api/portal/lenders":{"get":{"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["items"],"properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/Lender"}}}}}}}}}},"/api/portal/offers":{"get":{"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["items"],"properties":{"items":{"type":"array","items":{"$ref":"#/components/schemas/Offer"}}}}}}}}},"post":{"parameters":[{"$ref":"#/components/parameters/AuthHeader"},{"$ref":"#/components/parameters/ContentTypeHeader"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"400":{"$ref":"#/components/responses/Error"}}}},"/api/portal/readiness-leads":{"get":{"parameters":[{"$ref":"#/components/parameters/AuthHeader"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["items"],"properties":{"items":{"type":"array","items":{"type":"object","additionalProperties":true}}}}}}},"401":{"$ref":"#/components/responses/Error"}}}},"/api/portal/readiness-leads/{id}/convert":{"post":{"parameters":[{"$ref":"#/components/parameters/AuthHeader"},{"$ref":"#/components/parameters/ContentTypeHeader"}],"requestBody":{"required":false,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["applicationId"],"properties":{"applicationId":{"type":"string"}}}}}},"404":{"$ref":"#/components/responses/Error"}}}},"/api/portal/applications/{id}/readiness":{"get":{"parameters":[{"$ref":"#/components/parameters/AuthHeader"}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","required":["readinessLead"],"properties":{"readinessLead":{"type":"object","additionalProperties":true}}}}}},"404":{"$ref":"#/components/responses/Error"}}}},"/api/client/continuation/{token}":{"get":{"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"404":{"$ref":"#/components/responses/Error"}}}},"/api/portal/offers/{id}/status":{"patch":{"parameters":[{"$ref":"#/components/parameters/AuthHeader"},{"$ref":"#/components/parameters/ContentTypeHeader"}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"400":{"$ref":"#/components/responses/Error"}}}},"/api/banking/analysis":{"post":{"responses":{"200":{"description":"OK"}}}},"/api/credit-summary":{"post":{"responses":{"200":{"description":"OK"}}}},"/api/portal/lender-submissions":{"post":{"responses":{"201":{"description":"Created"}}}},"/api/messages":{"post":{"responses":{"201":{"description":"Created"}}}},"/api/calls/log":{"post":{"responses":{"201":{"description":"Created"}}}}},"components":{"parameters":{"AuthHeader":{"name":"Authorization","in":"header","required":true,"schema":{"type":"string","minLength":1}},"ContentTypeHeader":{"name":"Content-Type","in":"header","required":true,"schema":{"type":"string"}},"MultipartHeader":{"name":"Content-Type","in":"header","required":true,"schema":{"type":"string","pattern":"multipart/form-data"}}},"responses":{"Error":{"description":"Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}},"schemas":{"Application":{"$ref":"./schemas/application.json"},"Document":{"$ref":"./schemas/document.json"},"Lender":{"$ref":"./schemas/lender.json"},"Offer":{"$ref":"./schemas/offer.json"},"Error":{"$ref":"./schemas/error.json"}}}}